Text.Regex.TDFA.Pattern

Description

This Text.Regex.TDFA.Pattern module provides the Pattern data type and its subtypes. This Pattern type is used to represent the parsed form of a regular expression.

Documentation

data Pattern

Pattern is the type returned by the regular expression parser parseRegex. This is consumed by the Text.Regex.TDFA.CorePattern module and the tender leaves are nibbled by the Text.Regex.TDFA.TNFA module.

The DoPa field is the index of the component in the regex string r.

Constructors

PEmpty
(), matches the empty string.

PGroup (Maybe GroupIndex) Pattern
Group (r). Nothing indicates non-matching PGroup (never produced by parser parseRegex).

POr [Pattern]
Alternative r|s (flattened by starTrans).

PConcat [Pattern]
Sequence rs (flattened by starTrans).

PQuest Pattern
Zero or one repetitions r? (eliminated by starTrans).

PPlus Pattern
One or more repetitions r+ (eliminated by starTrans).

PStar Bool Pattern
Zero or more repetitions r*. True (default) means may accept the empty string on its first iteration.

PBound Int (Maybe Int) Pattern
Given number or repetitions r{n} or r{n,m} (eliminated by starTrans).

PCarat
^ matches beginning of input.

Fields

getDoPa :: DoPa

PDollar
$ matches end of input.

Fields

getDoPa :: DoPa

PDot
. matches any character.

Fields

getDoPa :: DoPa

PAny
Bracket expression [...].

Fields

getDoPa :: DoPa

getPatternSet :: PatternSet

PAnyNot
Inverted bracket expression [^...].

Fields

getDoPa :: DoPa

getPatternSet :: PatternSet

PEscape
Backslashed character c, may have special meaning.

Fields

getDoPa :: DoPa

getPatternChar :: Char

PChar
Single character, matches given character.

Fields

getDoPa :: DoPa

getPatternChar :: Char

PNonCapture Pattern
Tag for internal use, introduced by starTrans.

PNonEmpty Pattern
Tag for internal use, introduced by starTrans.

data PatternSet

Content of a bracket expression [...] organized into characters, POSIX character classes (e.g. [[:alnum:]]), collating elements (e.g. [.ch.], unused), and equivalence classes (e.g. [=a=], treated as characters).

Constructors

PatternSet (Maybe (Set Char)) (Maybe (Set PatternSetCharacterClass)) (Maybe (Set PatternSetCollatingElement)) (Maybe (Set PatternSetEquivalenceClass))

newtype PatternSetCharacterClass

Content of [: :], e.g. "alnum" for [:alnum:].

Constructors

PatternSetCharacterClass

Fields

unSCC :: String

newtype PatternSetCollatingElement

Content of [. .], e.g. "ch" for [.ch.].

Constructors

PatternSetCollatingElement

Fields

unSCE :: String

newtype PatternSetEquivalenceClass

Content of [= =], e.g. "a" for [=a=].

Constructors

PatternSetEquivalenceClass

Fields

unSEC :: String

type GroupIndex = Int

GroupIndex is for indexing submatches from capturing parenthesized groups (PGroup or Group).

newtype DoPa

Used to track elements of the pattern that accept characters or are anchors.

Constructors

DoPa

Fields

dopaIndex :: Int

decodeCharacterClass :: PatternSetCharacterClass -> String

This returns the strictly ascending list of characters represented by [: :] POSIX character classes. Unrecognized class names return an empty string.

Since: 1.3.2

decodePatternSet :: PatternSet -> Set Char

decodePatternSet cannot handle collating element and treats equivalence classes as just their definition and nothing more.

Since: 1.3.2

showPattern :: Pattern -> String

Internal use

starTrans :: Pattern -> Pattern

Do the transformation and simplification in a single traversal. This removes the PPlus, PQuest, and PBound values, changing to POr and PEmpty and PStar. For some PBound values it adds PNonEmpty and PNonCapture semantic marker. It also simplifies to flatten out nested POr and PConcat instances and eliminate some unneeded PEmpty values.

Internal use, operations to support debugging under ghci

starTrans' :: Pattern -> Pattern

simplify' :: Pattern -> Pattern

Function to transform a pattern into an equivalent, but less redundant form. Nested POr and PConcat are flattened. PEmpty is propagated.

dfsPattern

Arguments

:: (Pattern -> Pattern)
The transformation function.

-> Pattern
The Pattern to transform.

-> Pattern
The transformed Pattern.

Apply a Pattern transformation function depth first.